CMN-577: rename recommender to recommenders
