+++ base.scss
@@ -87,7 +87,7 @@
 @import "./src/stories/Library/Modals/modal-infomedia/infomedia";
 @import "./src/stories/Library/pincode-container/pincode-container";
 @import "./src/stories/Library/promo-bar/promo-bar";
-@import "./src/stories/Library/recommender/recommender";
+@import "./src/stories/Library/recommenders/recommenders";
 @import "./src/stories/Library/Lists/list-intermediates/list-intermediates";
 @import "./src/stories/Library/instant-loan/instant-loan";
 @import "./src/stories/Library/scroll-lock-background/scroll-lock-background";
@@ -97,7 +97,6 @@
 @import "./src/stories/Library/autosuggest-text/autosuggest-text";
 @import "./src/stories/Library/autosuggest-material/autosuggest-material";
 @import "./src/stories/Library/favorites-list-material-component/favorites-list-material-component";
-
 
 // Blocks
 @import "./src/stories/Blocks/footer/footer";
